Add homepage React card for each activity
Added navbar with activities and "Create Activity Button"
On the main page there are carts for each Activity with details and button for view in more detail. Clicking the View button opens a Detailed view with other details such as picture, a button to edit and a button to cancel.
Form comes with all informations that need to be  changed for editing an activity, also with submit and cancel button.

The Form does not have a functionality yet.