Make Form for Create/Edit Activity work locally
The form for Create new Activity/Edit existing activity now properly changes the activity locally in case of edit or creates a new activity and gives it proper GUID id . Both operations are done locally at the moment, no connection with the backend yet. Will future implement this relationship with the backend by calling the api with Axios.